씨젠/feature_cube/requirements.txt
numpy==1.21.4
tqdm==4.62.3
scikit-image==0.19.1
torch==1.10.1+cu113
torchvision==0.11.2+cu113
PyMySQL==1.0.2
pandas==1.3.5
Pillow==7.0.0